+++ editor/.eslint-plugin-local/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    'no-dispatch-use-effect': require('./no-dispatch-use-effect'),
+  },
+}
